requirements.txt
asgiref==3.11.1
Django==5.2.12
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
gunicorn==25.1.0
packaging==26.0
psycopg2-binary==2.9.11
python-dotenv==1.2.2
sqlparse==0.5.5
typing_extensions==4.15.0
